% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like_inference.R
\name{array_bic_aic}
\alias{array_bic_aic}
\title{Calculate the AIC and BIC.}
\usage{
array_bic_aic(sig_squared, p, mode_ident = NULL, mode_diag = NULL,
  mode_unstructured = NULL)
}
\arguments{
\item{sig_squared}{A numeric. The MLE of sigma^2 in the array normal model
(the 'variance' form of the total variation parameter).}

\item{p}{A vector of integers. The dimension of the data array (including
replication modes).}

\item{mode_ident}{A vector of integers. The modes assumed to have identity
covariances.}

\item{mode_diag}{A vector of integers. The modes assumed to have diagional
covariances.}

\item{mode_unstructured}{A vector of integers. The modes of assumed to have
unstructured covariances.}
}
\value{
\code{AIC} A numeric. The AIC of the model.

  \code{BIC} A numeric. The BIC of the model.
}
\description{
Calculate the AIC and BIC for Kronecker structured covariance models,
assuming the array normal distribution.
}
\details{
The AIC and BIC depend only on the data through the MLE of the total
variation parameter. Given this, the dimension of the array, and a
specification of which modes are the identity and which are unstructured,
this function will calculate the AIC and BIC.
}
\author{
David Gerard.
}

