% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff_functions.R
\name{Kom}
\alias{Kom}
\title{Commutation matrix.}
\usage{
Kom(m, n)
}
\arguments{
\item{m}{a natural number.}

\item{n}{another natural number.}
}
\description{
Construct the communtation matrix.
}
\details{
This function constructs the commutation matrix K_mn, which maps c(A) to
c(t(A)) for an m*n matrix A.
}
\examples{
m<-5 ; n<-4
A<-matrix(rnorm(m*n),m,n)
Kom(5,4)\%*\% c(A) - c(t(A))
}
\author{
Peter Hoff.
}
\keyword{algebra}

