% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tensor.R
\name{ldan}
\alias{ldan}
\title{Log-likelihood of array normal model.}
\usage{
ldan(E, Sig)
}
\arguments{
\item{E}{An array. This is the data.}

\item{Sig}{A list of symmetric positive definite matrices. These
are the component covariance matrices.}
}
\description{
\code{ldan} calculates the log-likelihood of the array normal
model, minus a constant.
}
\author{
David Gerard.
}

