% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tensor.R
\name{listprod}
\alias{listprod}
\title{Element-wise matrix products between two lists.}
\usage{
listprod(A, B)
}
\arguments{
\item{A}{A list of matrices.}

\item{B}{A second list of matrices.}
}
\value{
A list \code{C} such that \code{C[[i]] = A[[i]] \%*\% B[[i]]}.
}
\description{
Given two lists of matrices with conformable dimensions,
\code{listprod} returns a list whose elements are the matrix
products of the elements of these two lists.
}
\author{
David Gerard.
}

