% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff_functions.R
\name{rsan}
\alias{rsan}
\title{Standard normal array.}
\usage{
rsan(dim)
}
\arguments{
\item{dim}{a vector of positive integers.}
}
\description{
Generate an array of iid standard normal variables.
}
\details{
This functions generates an array of dimension \code{dim} filled
with iid standard normal variables.
}
\examples{
tensr:::rsan(c(5,4,3))
}
\author{
Peter Hoff.
}
\keyword{multivariate}
\keyword{simulation}

