% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensr.R
\docType{package}
\name{tensr-package}
\alias{tensr}
\alias{tensr-package}
\title{tensr: Covariance Inference and Decompositions for Tensor Datasets}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of functions for Kronecker structured covariance estimation and testing under the array normal model. For estimation, maximum likelihood and Bayesian equivariant estimation procedures are implemented. For testing, a likelihood ratio testing procedure is available. This package also contains additional functions for manipulating and decomposing tensor data sets. This work was partially supported by NSF grant DMS-1505136. Details of the methods are described in Gerard and Hoff (2015) \doi{10.1016/j.jmva.2015.01.020} and Gerard and Hoff (2016) \doi{10.1016/j.laa.2016.04.033}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dcgerard/tensr}
  \item Report bugs at \url{https://github.com/dcgerard/tensr/issues}
}

}
\author{
\strong{Maintainer}: David Gerard \email{gerard.1787@gmail.com} (\href{https://orcid.org/0000-0001-9450-5023}{ORCID})

Authors:
\itemize{
  \item Peter Hoff \email{peter.hoff@duke.edu}
}

}
\keyword{internal}
