% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff_functions.R
\name{topK}
\alias{topK}
\title{Top K elements of a vector.}
\usage{
topK(x, K = 1, ignoreties = TRUE)
}
\arguments{
\item{x}{The vector.}

\item{K}{The number of indices to return.}

\item{ignoreties}{If \code{FALSE}, will return a vector of the
indices whose elements are greater than or equal to the Kth
largest element, resulting in a vector possibly of length
greater than \code{K} in the case of ties.}
}
\description{
Identify top K elements of a vector.
}
\details{
This function returns the indices corresponding to the top elements
of a vector.
}
\examples{
x <- c(3, 6, 2, 4, 1)
tensr:::topK(x, 3)
}
\author{
Peter Hoff.
}
