% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tensor.R
\name{tr}
\alias{tr}
\title{Trace of a matrix.}
\usage{
tr(X)
}
\arguments{
\item{X}{A matrix whose diagonal elements will be added together.}
}
\value{
The sum of the diagonal elements of X.
}
\description{
Returns the sum of the diagonal elements of a matrix.
}
\details{
This returns the trace of a matrix, which is just the sum of its
diagonal elements.
}
\examples{
X <- matrix(1:4, nrow = 2, ncol = 2)
X
tr(X)
}
\author{
Peter Hoff.
}
