% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff_functions.R
\name{tsum}
\alias{tsum}
\title{Tucker sum.}
\usage{
tsum(X, A)
}
\arguments{
\item{X}{A real array.}

\item{A}{A list of real matrices.}
}
\description{
Computes the Tucker sum of an array and a list of matrices.
}
