% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoff_functions.R
\name{anorm_cd}
\alias{anorm_cd}
\title{Array normal conditional distributions.}
\usage{
anorm_cd(Y, M, S, saidx)
}
\arguments{
\item{Y}{a real valued array}

\item{M}{mean of \code{Y}}

\item{S}{list of mode-specific covariance matrices of \code{Y}}

\item{saidx}{list of indices for which the conditional mean and variance
should be computed}
}
\description{
Conditional mean and variance of a subarray.
}
\details{
This function calculates the conditional mean and variance in the array
normal model.
}
\examples{
m<-c(6,5,4)
K<-length(m)
M<-rsan(m)
S<-list() ; for(k in 1:K) { S[[k]]<-rwish(diag(m[k]))/m[k] }
Y<-M+atrans(rsan(m),lapply(S,mhalf))

SY<-Y*0
SS<-list() ; for(k in 1:K) { SS[[k]]<-matrix(0,m[k],m[k]) }
for(s in 1:1e4)
{
  saidx<-list()
  for(k in 1:K){ saidx[[k]]<-sort(sample(1:m[k],1+rbinom(1,m[k]-1,.5))) }
  MS<-anorm_cd(Y,M,S,saidx)
  Ya<- MS$Mab + atrans(rsan(sapply(saidx,length)),lapply(MS$Sab,mhalf))
  Y[arrIndices(saidx)]<-Ya
  SY<-SY+Y
  for(k in 1:K){SS[[k]]<-SS[[k]]+mat(Y-M,k)\%*\%t(mat(Y-M,k)) }
}
par(mfrow=c(2,2))
plot(M,SY/s)  ; abline(0,1)
for (k in 1:K) {plot(S[[k]]/tr(S[[k]]),SS[[k]]/tr(SS[[k]])) ; abline(0,1) }
}
\author{
Peter Hoff.
}
\keyword{arrays}
\keyword{multivariate}

